List All Files in a Directory

Lists all files in a specified directory using base::list.files(), with options to include subdirectories and hidden files. If no files are found, provides informative error messages.

Usage

list_files(path, recursive, all.files)

Arguments

path

a character vector of full path names; the default corresponds to the working directory, getwd(). Tilde expansion (see path.expand) is performed. Missing values will be ignored. Elements with a marked encoding will be converted to the native encoding (and if that fails, considered non-existent).

recursive

logical. Should the listing recurse into directories?

all.files

a logical value. If FALSE, only the names of visible files are returned (following Unix-style visibility, that is files whose name does not start with a dot). If TRUE, all file names will be returned.

Value

A character vector of normalized file paths.